added fonts and reorganized a lil
