+++ src/App.js
@@ -9,14 +9,6 @@ function App() {
         <p>
           Indigenous Interpreters
         </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
       </header>
     </div>
   );
