+++ src/components/Navigation/Navigation.css
@@ -1,0 +1,4 @@
+.Navigation li a:hover:not(.active) {
+    color: black;
+    border-bottom: 2px solid #03A0B5;
+  }
