+++ src/containers/AdminPage/Admin.js
@@ -2,4 +2,3 @@ import React from 'react';
 import classes from './Admin.css'
 
 
-
